leaflet client: improvements to session management for Calltaker/Fieldtrip modules

+++ opentripplanner-datastore/app/controllers/FieldTrip.java
@@ -7,6 +7,8 @@ import models.fieldtrip.GroupItinerary;
 import models.fieldtrip.GTFSTrip;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import static controllers.Application.checkLogin;
+import static controllers.Calltaker.checkAccess;
 import play.*;
 import play.mvc.*;
 
@@ -19,7 +21,19 @@ import play.data.binding.As;
 
 
 public class FieldTrip extends Application {
-
+    
+    @Util
+    public static void checkAccess(TrinetUser user) {
+        if(user == null) {
+            System.out.println("null user in FieldTrip module");
+            forbidden("null user");
+        }
+        if(!user.hasFieldTripAccess()) {
+            System.out.println("User " + user.username + " has insufficient access for FieldTrip module");
+            forbidden("insufficient access privileges");
+        }
+    }
+    
     /*@Before(unless={"newTrip","addTripFeedback","getCalendar","newRequest","newRequestForm"}, priority=1)
     public static void checkLogin () {
         String username = params.get("userName");
@@ -61,6 +75,9 @@ public class FieldTrip extends Application {
     }
 
     public static void getFieldTrip(long id) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         ScheduledFieldTrip fieldTrip = ScheduledFieldTrip.findById(id);
         Gson gson = new GsonBuilder()
           .excludeFieldsWithoutExposeAnnotation()  
@@ -69,6 +86,9 @@ public class FieldTrip extends Application {
     }
     
     public static void getFieldTrips(@As("MM/dd/yyyy") Date date, Integer limit) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         System.out.println("getFTs, date="+date);
         List<ScheduledFieldTrip> trips;
         String sql = "";
@@ -94,6 +114,9 @@ public class FieldTrip extends Application {
     }
     
     public static void getGTFSTripsInUse(@As("MM/dd/yyyy") Date date, Integer limit) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         System.out.println("getFTs, date="+date);
         List<ScheduledFieldTrip> trips;
         String sql = "";
@@ -127,6 +150,9 @@ public class FieldTrip extends Application {
 
     
     public static void newTrip(long requestId, ScheduledFieldTrip trip) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         FieldTripRequest ftRequest = FieldTripRequest.findById(requestId);
 
         //TODO: is setting id to null the right way to ensure that an
@@ -134,12 +160,8 @@ public class FieldTrip extends Application {
         trip.id = null;
         trip.request = ftRequest;
         trip.serviceDay = trip.departure;
-        //trip.timeStamp = new Date();
-        /*User user = getUser();
-        if (!user.canScheduleFieldTrips()) {
-            //TODO: is this safe if those itineraries exist?
-            trip.groupItineraries.clear();
-        }*/
+        trip.createdBy = user.username;
+        
         trip.save();
         
         ftRequest.trips.add(trip);
@@ -151,6 +173,9 @@ public class FieldTrip extends Application {
     }
     
     public static void addItinerary(long fieldTripId, GroupItinerary itinerary, GTFSTrip[] trips) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         System.out.println("aI / fieldTripId="+fieldTripId);
         ScheduledFieldTrip fieldTrip = ScheduledFieldTrip.findById(fieldTripId);
         //System.out.println("aI / fieldTrip="+fieldTrip);
@@ -170,6 +195,9 @@ public class FieldTrip extends Application {
     }
 
     public static void deleteTrip(Long id) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         ScheduledFieldTrip trip = ScheduledFieldTrip.findById(id);
         trip.delete();
         renderJSON(id);
@@ -196,6 +224,9 @@ public class FieldTrip extends Application {
     }
     
     public static void getRequests(Integer limit) {
+        TrinetUser user = checkLogin();        
+        checkAccess(user);
+      
         List<FieldTripRequest> requests;
         String sql = "order by timeStamp desc";
         if(limit == null)
@@ -211,28 +242,6 @@ public class FieldTrip extends Application {
         renderJSON(gson.toJson(requests));
     }
     
-    
-    /*public static void setInboundTrip(long requestId, long tripId) {
-        FieldTripRequest ftRequest = FieldTripRequest.findById(requestId);
-        ScheduledFieldTrip trip = ScheduledFieldTrip.findById(tripId);
-        ftRequest.trips.add(trip);
-        trip.request = ftRequest;
-        trip.requestOrder = 1;
-        trip.save();
-        ftRequest.save();
-        renderJSON(requestId);
-    }
-
-    public static void setOutboundTrip(long requestId, long tripId) {
-        FieldTripRequest ftRequest = FieldTripRequest.findById(requestId);
-        ScheduledFieldTrip trip = ScheduledFieldTrip.findById(tripId);
-        ftRequest.trips.add(trip);
-        trip.request = ftRequest;
-        trip.requestOrder = 0;
-        trip.save();
-        ftRequest.save();
-        renderJSON(requestId);
-    }*/
 
     /* FieldTripFeedback */
